+++ src/gitInfo.ts
@@ -1,5 +1,5 @@
 export const GIT_INFO = {
   branch: "fix/ANT-2760_replace_rds_component_into_customize_one",
-  commit: "5437e65",
-  commitTime: "2025-06-17T12:59:59.825Z"
+  commit: "5725eab",
+  commitTime: "2025-06-20T08:36:27.358Z"
 };
