--- src/shared/types/Tailwind.type.ts
@@ -1,42 +1,0 @@
-import { TailwindColorClass } from '@/shared/types/TailwindColorClass.type.ts';
-
-export type ZIndex =
-  | 'z-auto'
-  | 'z-0'
-  | 'z-10'
-  | 'z-20'
-  | 'z-30'
-  | 'z-40'
-  | 'z-50'
-  | '-z-10'
-  | '-z-30'
-  | '-z-40'
-  | '-z-20'
-  | '-z-50';
-
-type PseudoClass =
-  | ''
-  | 'hover:'
-  | 'active:'
-  | 'focus:'
-  | 'focus-visible:'
-  | '[&]:'
-  | 'after:'
-  | 'before:'
-  | '[&.active]:';
-
-type TailwindPrefix = 'text' | 'bg' | 'border' | 'outline';
-
-type Delay = 0 | 75 | 100 | 150 | 200 | 300 | 500 | 700 | 1000;
-export type TailwindDelayClass = `delay-${Delay}`;
-
-type Duration = 0 | 75 | 100 | 150 | 200 | 300 | 500 | 700 | 1000;
-export type TailwindDurationClass = `duration-${Duration}`;
-
-type Rotate = 0 | 1 | 2 | 3 | 6 | 12 | 45 | 90 | 180;
-export type TailwindRotateClass = `rotate-${Rotate}`;
-
-export type TailwindUtilityColorClass<
-  Prefix extends TailwindPrefix,
-  TPseudoClass extends PseudoClass = '',
-> = `${TPseudoClass}${Prefix}-${TailwindColorClass}`;
