+++ src/shared/utils/tailwindClass.ts
@@ -1,5 +1,4 @@
-import { PseudoClass, TailwindPrefix } from '../types/Tailwind.type';
-import { TailwindColorClass } from '@/shared/types';
+import { PseudoClass, TailwindColorClass, TailwindPrefix } from '@/shared/types';
 
 export const buildColorClass = (prefix: TailwindPrefix, color: TailwindColorClass, pseudoClass?: PseudoClass) =>
   `${pseudoClass || ''}${prefix}-${color}`;
